structure the bot file structures
