+++ bin/main.rb
@@ -1,0 +1,3 @@
+require 'slack-ruby-bot'
+require_relative '../app/app.rb'
+require_relative '../app/lib/bot.rb'
