+++ config.ru
@@ -1,10 +1,8 @@
-# frozen_string_literal: true
-
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 
 require 'dotenv'
 Dotenv.load
 
-require './main.rb'
+require './bin/main.rb'
 
 Bafiambot::Bot.run
