--- main.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require 'slack-ruby-bot'
-require './app/app.rb'
-require './app/commands/bot.rb'
