+++ spec/bot_spec.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 require './spec/spec_helper.rb'
 
 describe Bafiambot::Bot do
