+++ spec/spec_helper.rb
@@ -1,6 +1,4 @@
-# frozen_string_literal: true
-
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..'))
 
 require 'slack-ruby-bot/rspec'
-require 'main.rb'
+require_relative '../bin/main.rb'
